+++ server/handlers/tokenHandler/index.js
@@ -1,17 +1,18 @@
 const TokenService = require('../../services/TokenService');
-const { tokenGetSchema, tokenGetTransactionsByIdSchema } = require('./schemas');
+const { tokenGetSchema, tokenIdSchema, tokenGetTransactionsByIdSchema } = require('./schemas');
 
 const tokenGet = async (req, res) => {
-  await tokenGetSchema.validateAsync(req.query, { abortEarly: false });
+  const validatedQuery = await tokenGetSchema.validateAsync(req.query, { abortEarly: false });
 
-  const { limit, wallet, offset } = req.query;
+  const { limit, wallet, offset } = validatedQuery;
+  const { wallet_id } = req
   const tokenService = new TokenService();
 
   const tokens = await tokenService.getTokens({
     wallet,
     limit,
     offset,
-    walletLoginId: req.wallet_id,
+    walletLoginId: wallet_id,
   });
 
   res.status(200).json({
@@ -20,11 +21,15 @@ const tokenGet = async (req, res) => {
 };
 
 const tokenGetById = async (req, res) => {
-  const { id } = req.params;
+  const validatedParams = await tokenIdSchema.validateAsync(req.params, {
+    abortEarly: false,
+  });
+  const {id} = validatedParams;
+  const {wallet_id} = req
   const tokenService = new TokenService();
   const token = await tokenService.getById({
     id,
-    walletLoginId: req.wallet_id,
+    walletLoginId: wallet_id,
   });
 
   res.status(200).json(token);
@@ -32,17 +37,21 @@ const tokenGetById = async (req, res) => {
 
 const tokenGetTransactionsById = async (req, res) => {
   // validate input
-  await tokenGetTransactionsByIdSchema.validateAsync(req.query, {
+  const validatedQuery = await tokenGetTransactionsByIdSchema.validateAsync(req.query, {
     abortEarly: false,
   });
-  const { limit, offset } = req.query;
-  const { id } = req.params;
+  const { limit, offset } = validatedQuery;
+  const validatedParams = await tokenIdSchema.validateAsync(req.params, {
+    abortEarly: false,
+  });
+  const { id } = validatedParams;
+  const {wallet_id} = req
   const tokenService = new TokenService();
   const transactions = await tokenService.getTransactions({
     tokenId: id,
     limit,
     offset,
-    walletLoginId: req.wallet_id,
+    walletLoginId: wallet_id,
   });
   res.status(200).json({
     history: transactions,