+++ server/handlers/transferHandler/index.js
@@ -13,9 +13,10 @@ const transferPost = async (req, res) => {
   });
   const transferService = new TransferService();
 
+  const { wallet_id } = req;
   const { result, status } = await transferService.initiateTransfer(
-    validatedData,
-    req.wallet_id,
+    validatedBody,
+    wallet_id,
   );
 
   const modifiedTransfer = {
@@ -29,65 +30,80 @@ const transferPost = async (req, res) => {
 };
 
 const transferIdAcceptPost = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-
-  const transferService = new TransferService();
-  const result = await transferService.acceptTransfer(
-    req.params.transfer_id,
-    req.wallet_id,
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
   );
+
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
+  const transferService = new TransferService();
+  const result = await transferService.acceptTransfer(transfer_id, wallet_id);
 
   res.json(result);
 };
 
 const transferIdDeclinePost = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-
-  const transferService = new TransferService();
-  const result = await transferService.declineTransfer(
-    req.params.transfer_id,
-    req.wallet_id,
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
   );
+
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
+  const transferService = new TransferService();
+  const result = await transferService.declineTransfer(transfer_id, wallet_id);
 
   res.json(result);
 };
 
 const transferIdDelete = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-
-  const transferService = new TransferService();
-  const result = await transferService.cancelTransfer(
-    req.params.transfer_id,
-    req.wallet_id,
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
   );
+
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
+  const transferService = new TransferService();
+  const result = await transferService.cancelTransfer(transfer_id, wallet_id);
 
   res.json(result);
 };
 
 const transferIdFulfill = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-  await transferIdFulfillSchema.validateAsync(req.body, { abortEarly: false });
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
+  );
+  const validatedBody = await transferIdFulfillSchema.validateAsync(req.body, {
+    abortEarly: false,
+  });
 
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
   const transferService = new TransferService();
-
   const result = await transferService.fulfillTransfer(
-    req.wallet_id,
-    req.params.transfer_id,
-    req.body,
+    wallet_id,
+    transfer_id,
+    validatedBody,
   );
   res.json(result);
 };
 
 const transferGet = async (req, res) => {
-  await transferGetQuerySchema.validateAsync(req.query, { abortEarly: false });
+  const validatedQuery = await transferGetQuerySchema.validateAsync(req.query, {
+    abortEarly: false,
+  });
+
+  const { limit, offset, ...params } = validatedQuery;
+  const { wallet_id } = req;
 
   const transferService = new TransferService();
 
-  const { limit = 200, offset = 0, ...params } = req.query;
-
   const { transfers, count } = await transferService.getByFilter(
     { ...params, limit, offset },
-    req.wallet_id,
+    wallet_id,
   );
 
   const modifiedTransfers = transfers.map((t) => ({
@@ -104,13 +120,15 @@ const transferGet = async (req, res) => {
 };
 
 const transferIdGet = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-
-  const transferService = new TransferService();
-  const result = await transferService.getTransferById(
-    req.params.transfer_id,
-    req.wallet_id,
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
   );
+
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
+  const transferService = new TransferService();
+  const result = await transferService.getTransferById(transfer_id, wallet_id);
 
   const modifiedTransfer = {
     ...result,
@@ -123,17 +141,24 @@ const transferIdGet = async (req, res) => {
 };
 
 const transferIdTokenGet = async (req, res) => {
-  await transferIdParamSchema.validateAsync(req.params, { abortEarly: false });
-  await transferLimitOffsetQuerySchema.validateAsync(req.query, {
-    abortEarly: false,
-  });
+  const validatedParams = await transferIdParamSchema.validateAsync(
+    req.params,
+    { abortEarly: false },
+  );
+  const validatedQuery = await transferLimitOffsetQuerySchema.validateAsync(
+    req.query,
+    {
+      abortEarly: false,
+    },
+  );
 
-  const { limit, offset } = req.query;
-
+  const { limit, offset } = validatedQuery;
+  const { transfer_id } = validatedParams;
+  const { wallet_id } = req;
   const transferService = new TransferService();
   const tokens = await transferService.getTokensByTransferId(
-    req.params.transfer_id,
-    req.wallet_id,
+    transfer_id,
+    wallet_id,
     limit,
     offset,
   );