+++ server/services/AuthService.spec.js
@@ -8,10 +8,10 @@ const Event = require('../models/Event');
 
 describe('AuthService', () => {
   it('signin', async () => {
-    const walletObject = { salt: 'salt', password: 'hash', id: 'id' };
-    const getByIdOrNameStub = Sinon.stub(
+    const walletObject = { salt: 'salt', password: 'hash' };
+    const getByNameStub = Sinon.stub(
       WalletService.prototype,
-      'getByIdOrName',
+      'getByName',
     ).resolves(walletObject);
 
     const logEventStub = Sinon.stub(Event.prototype, 'logEvent');
@@ -19,20 +19,13 @@ describe('AuthService', () => {
     const jwtSignStub = Sinon.stub(JWTService, 'sign').resolves('token');
     const details = { wallet: 'wallet', password: 'password' };
     const token = await AuthService.signIn(details);
-    expect(getByIdOrNameStub.calledOnceWithExactly(details.wallet)).eql(true);
+    expect(getByNameStub.calledOnceWithExactly(details.wallet)).eql(true);
     expect(sha512Stub.calledOnceWithExactly(details.password, 'salt')).eql(
       true,
     );
     expect(jwtSignStub.calledOnceWithExactly(walletObject)).eql(true);
     expect(token).eql('token');
-    expect(
-      logEventStub.calledOnceWithExactly({
-        wallet_id: 'id',
-        type: 'login',
-        payload: {},
-      }),
-    ).to.eql(true);
-    getByIdOrNameStub.restore();
+    getByNameStub.restore();
     sha512Stub.restore();
     jwtSignStub.restore();
     logEventStub.restore();
@@ -40,23 +33,22 @@ describe('AuthService', () => {
 
   it('failed signin', async () => {
     const walletObject = { salt: 'salt', password: 'password' };
-    const getByIdOrNameStub = Sinon.stub(
+    const getByNameStub = Sinon.stub(
       WalletService.prototype,
-      'getByIdOrName',
+      'getByName',
     ).resolves(walletObject);
     const logEventStub = Sinon.stub(Event.prototype, 'logEvent');
     const sha512Stub = Sinon.stub(HashService, 'sha512').returns('hash');
     const jwtSignStub = Sinon.stub(JWTService, 'sign').resolves('token');
     const details = { wallet: 'wallet', password: 'password' };
     const token = await AuthService.signIn(details);
-    expect(getByIdOrNameStub.calledOnceWithExactly(details.wallet)).eql(true);
+    expect(getByNameStub.calledOnceWithExactly(details.wallet)).eql(true);
     expect(sha512Stub.calledOnceWithExactly(details.password, 'salt')).eql(
       true,
     );
     expect(jwtSignStub.notCalled).eql(true);
     expect(token).eql(false);
-    expect(logEventStub.notCalled).to.eql(true);
-    getByIdOrNameStub.restore();
+    getByNameStub.restore();
     sha512Stub.restore();
     jwtSignStub.restore();
     logEventStub.restore();