+++ server/services/WalletService.spec.js
@@ -210,7 +210,7 @@ describe('WalletService', () => {
     const walletId2 = uuid.v4();
     const limitOptions = {
       limit: 10,
-      offet: 0,
+      offset: 0,
     };
     const result = [
       {
@@ -239,11 +239,11 @@ describe('WalletService', () => {
         id,
         limitOptions,
         'name',
+         false,
       );
-
       expect(
-        getAllWalletsStub.calledOnceWithExactly(id, limitOptions, 'name', true),
-      ).eql(true);
+        getAllWalletsStub).calledOnceWithExactly(id, limitOptions, 'name', true)
+    // ).eql(true);
       expect(allWallets).eql({ wallets: result, count });
     });